RFC 9022 - CiSpecificsLocationItemView v1 fertig
git-svn-id: http://10.205.116.236/svn/AIR@297 d2170373-ae6f-44b7-9150-2eec3990abfc

+++ branches/2.00/src/com/bayerbbs/applrepos/dto/CiSupportStuffDTO.java
@@ -8,7 +8,7 @@ public class CiSupportStuffDTO implements Serializable {
 	
 	private Long ciSupportStuffId;
 	private Long ciSupportStuffTypeId;
-	private Long tableId;
+	private Integer tableId;
 	private Long ciId;
 	private String ciSupportStuffValue;
 
@@ -24,10 +24,10 @@ public class CiSupportStuffDTO implements Serializable {
 	public void setCiSupportStuffTypeId(Long ciSupportStuffTypeId) {
 		this.ciSupportStuffTypeId = ciSupportStuffTypeId;
 	}
-	public Long getTableId() {
+	public Integer getTableId() {
 		return tableId;
 	}
-	public void setTableId(Long tableId) {
+	public void setTableId(Integer tableId) {
 		this.tableId = tableId;
 	}
 	public Long getCiId() {
